requirements.txt
taipy
openai
ollama
aisuite
langchain_ollama
google-genai
bs4